Refactorization round 1 completed
Highlight searched words, include more columns for search

Highlight searched words, include more columns for search

Removed empty file and console log statements

Fixed issue around infinite scroll and searching

Removed commented console logs

Fixed number hyperlinking issue
